requirements.txt
fastapi
uvicorn[standard]
pytesseract
pdf2image
opencv-python
pillow
python-multipart
python-dotenv
spacy
sqlalchemy
PyPDF2
python-dateutil
